+++ src/apiflask/schemas.py
@@ -1,8 +1,9 @@
 import typing as t
 
-from marshmallow import Schema as Schema
+from marshmallow import Schema as BaseSchema
 from marshmallow.fields import Integer
 from marshmallow.fields import URL
+from marshmallow.orderedset import OrderedSet
 
 
 # schema for the detail object of validation error response
@@ -48,6 +49,19 @@ http_error_schema: t.Dict[str, t.Any] = {
     },
     'type': 'object'
 }
+
+
+class Schema(BaseSchema):
+    """A base schema for all schemas.
+
+    The different between marshmallow's `Schema` and APIFlask's `Schema` is that the latter
+    sets `set_class` to `OrderedSet` by default.
+
+    *Version Added: 1.2.0*
+    """
+    # use ordered set to keep the order of fields
+    # can be removed when https://github.com/marshmallow-code/marshmallow/pull/1896 is merged
+    set_class = OrderedSet
 
 
 class EmptySchema(Schema):