+++ tests/test_decorator_doc.py
@@ -1,8 +1,8 @@
 import pytest
-from flask.views import MethodView
 from openapi_spec_validator import validate_spec
 
 from .schemas import Foo
+from apiflask.views import MethodView
 
 
 def test_doc_summary_and_description(app, client):
@@ -194,7 +194,7 @@ def test_doc_responses(app, client):
     assert rv.json['paths']['/bar']['get']['responses'][
         '200']['description'] == 'Successful response'
     assert rv.json['paths']['/bar']['get']['responses'][
-        '400']['description'] == 'Validation error'
+        '400']['description'] == 'Bad Request'
     assert '404' in rv.json['paths']['/foo']['get']['responses']
     assert rv.json['paths']['/bar']['get']['responses'][
         '404']['description'] == 'Not Found'
@@ -242,7 +242,7 @@ def test_doc_responses_with_methodview(app, client):
     assert rv.json['paths']['/bar']['get']['responses'][
         '200']['description'] == 'Successful response'
     assert rv.json['paths']['/bar']['get']['responses'][
-        '400']['description'] == 'Validation error'
+        '400']['description'] == 'Bad Request'
     assert '404' in rv.json['paths']['/foo']['get']['responses']
     assert rv.json['paths']['/bar']['get']['responses'][
         '404']['description'] == 'Not Found'