+++ tests/test_decorator_input.py
@@ -1,17 +1,18 @@
 import io
 
 import pytest
-from flask.views import MethodView
 from openapi_spec_validator import validate_spec
 from werkzeug.datastructures import FileStorage
 
 from .schemas import Bar
+from .schemas import EnumPathParameter
 from .schemas import Files
 from .schemas import Foo
 from .schemas import Form
 from .schemas import FormAndFiles
 from .schemas import Query
 from apiflask.fields import String
+from apiflask.views import MethodView
 
 
 def test_input(app, client):
@@ -28,7 +29,7 @@ def test_input(app, client):
 
     for rule in ['/foo', '/bar']:
         rv = client.post(rule)
-        assert rv.status_code == 400
+        assert rv.status_code == 422
         assert rv.json == {
             'detail': {
                 'json': {'name': ['Missing data for required field.']}
@@ -37,7 +38,7 @@ def test_input(app, client):
         }
 
         rv = client.post(rule, json={'id': 1})
-        assert rv.status_code == 400
+        assert rv.status_code == 422
         assert rv.json == {
             'detail': {
                 'json': {'name': ['Missing data for required field.']}
@@ -68,7 +69,7 @@ def test_input_with_query_location(app, client):
         return {'name': schema['name'], 'name2': schema2['name2']}
 
     rv = client.post('/foo')
-    assert rv.status_code == 400
+    assert rv.status_code == 422
     assert rv.json == {
         'detail': {
             'query': {'name': ['Missing data for required field.']}
@@ -77,7 +78,7 @@ def test_input_with_query_location(app, client):
     }
 
     rv = client.post('/foo?id=1&name=bar')
-    assert rv.status_code == 400
+    assert rv.status_code == 422
     assert rv.json == {
         'detail': {
             'query': {'name2': ['Missing data for required field.']}
@@ -181,6 +182,38 @@ def test_input_with_form_and_files_location(app, client):
     assert rv.json == {'name': True, 'image': True}
 
 
+def test_input_with_path_location(app, client):
+    @app.get('/<image_type>')
+    @app.input(EnumPathParameter, location='path')
+    def index(image_type, _):
+        return {'image_type': image_type}
+
+    rv = client.get('/openapi.json')
+    assert rv.status_code == 200
+    # TODO: Failed validating 'oneOf' in schema
+    # https://github.com/p1c2u/openapi-spec-validator/issues/113
+    # validate_spec(rv.json)
+    assert '/{image_type}' in rv.json['paths']
+    assert len(rv.json['paths']['/{image_type}']['get']['parameters']) == 1
+    assert rv.json['paths']['/{image_type}']['get']['parameters'][0]['in'] == 'path'
+    assert rv.json['paths']['/{image_type}']['get']['parameters'][0]['name'] == 'image_type'
+    assert rv.json['paths']['/{image_type}']['get']['parameters'][0]['schema'] == {
+        'type': 'string',
+        'enum': ['jpg', 'png', 'tiff', 'webp'],
+    }
+
+    rv = client.get('/png')
+    assert rv.status_code == 200
+    assert rv.json == {'image_type': 'png'}
+
+    rv = client.get('/gif')
+    assert rv.status_code == 422
+    assert rv.json['message'] == 'Validation error'
+    assert 'path' in rv.json['detail']
+    assert 'image_type' in rv.json['detail']['path']
+    assert rv.json['detail']['path']['image_type'] == ['Must be one of: jpg, png, tiff, webp.']
+
+
 @pytest.mark.parametrize('locations', [
     ['files', 'form'],
     ['files', 'json'],
@@ -236,7 +269,7 @@ def test_input_with_dict_schema(app, client):
         return body
 
     rv = client.get('/foo')
-    assert rv.status_code == 400
+    assert rv.status_code == 422
     assert rv.json == {
         'detail': {
             'query': {'name': ['Missing data for required field.']}
@@ -249,7 +282,7 @@ def test_input_with_dict_schema(app, client):
     assert rv.json == {'name': 'grey'}
 
     rv = client.post('/bar')
-    assert rv.status_code == 400
+    assert rv.status_code == 422
     assert rv.json == {
         'detail': {
             'json': {'name': ['Missing data for required field.']}